requirements.txt
requests>=2.31.0
psycopg2-binary>=2.9.5
pandas>=2.0.0
APScheduler>=3.10.0
python-dotenv>=1.0.0
flask>=2.3.0
plotly>=5.15.0
gunicorn>=21.2.0